Move state reducer to it's own file
